requirements.txt
Pillow==9.2.0
opencv-python==4.5.3.56
cv2